feat(2-registries): introduce the concept of virtual directory
