+++ crates/weaver_codegen_test/build.rs
@@ -12,6 +12,7 @@ use std::io::Write;
 use std::path::{Component, Path, PathBuf};
 use std::process::exit;
 use weaver_common::in_memory::LogMessage;
+use weaver_common::vdir::VirtualDirectoryPath;
 use weaver_common::{in_memory, Logger};
 use weaver_forge::config::{Params, WeaverConfig};
 use weaver_forge::file_loader::FileSystemFileLoader;
@@ -19,7 +20,6 @@ use weaver_forge::registry::ResolvedRegistry;
 use weaver_forge::{OutputDirective, TemplateEngine};
 use weaver_resolver::SchemaResolver;
 use weaver_semconv::registry::SemConvRegistry;
-use weaver_semconv::registry_path::RegistryPath;
 use weaver_semconv::registry_repo::RegistryRepo;
 
 const SEMCONV_REGISTRY_PATH: &str = "./semconv_registry/";
@@ -40,7 +40,7 @@ fn main() {
     let logger = in_memory::Logger::new(0);
 
     // Load and resolve the semantic convention registry
-    let registry_path = RegistryPath::LocalFolder {
+    let registry_path = VirtualDirectoryPath::LocalFolder {
         path: SEMCONV_REGISTRY_PATH.into(),
     };
     let registry_repo =
