+++ crates/weaver_common/Cargo.toml
@@ -18,7 +18,25 @@ serde_json.workspace = true
 miette.workspace = true
 thiserror.workspace = true
 rouille.workspace = true
+regex.workspace = true
+once_cell.workspace = true
+schemars.workspace = true
+ureq.workspace = true
+
+tempdir = "0.3.7"
+dirs = "6.0.0"
+gix = { version = "0.70.0", default-features = false, features = [
+    "comfort",
+    "blocking-http-transport-reqwest",
+    "max-performance-safe",
+    "worktree-mutation",
+    "blocking-http-transport-reqwest-rust-tls",
+] }
+flate2 = "1.0.35"
+tar = "0.4.43"
+zip = "2.2.2"
 
 [dev-dependencies]
 ureq.workspace = true
+walkdir.workspace = true
