+++ crates/weaver_semconv/src/lib.rs
@@ -20,7 +20,6 @@ pub mod manifest;
 pub mod metric;
 pub mod provenance;
 pub mod registry;
-pub mod registry_path;
 pub mod registry_repo;
 pub mod semconv;
 pub mod stability;
@@ -271,50 +270,9 @@ pub enum Error {
         error: String,
     },
 
-    /// Home directory not found.
-    #[error("Home directory not found")]
-    HomeDirNotFound,
-
-    /// Cache directory not created.
-    #[error("Cache directory not created: {message}")]
-    CacheDirNotCreated {
-        /// The error message
-        message: String,
-    },
-
-    /// Git repo not created.
-    #[error("Git repo `{repo_url}` not created: {message}")]
-    GitRepoNotCreated {
-        /// The git repo URL
-        repo_url: String,
-        /// The error message
-        message: String,
-    },
-
-    /// A git error occurred.
-    #[error("Git error occurred while cloning `{repo_url}`: {message}")]
-    GitError {
-        /// The git repo URL
-        repo_url: String,
-        /// The error message
-        message: String,
-    },
-
-    /// An invalid registry path.
-    #[error("The registry path `{path}` is invalid: {error}")]
-    InvalidRegistryPath {
-        /// The registry path
-        path: String,
-        /// The error message
-        error: String,
-    },
-
-    /// An invalid registry archive.
-    #[error("This archive `{archive}` is not supported. Supported formats are: .tar.gz, .zip")]
-    UnsupportedRegistryArchive {
-        /// The registry archive path
-        archive: String,
-    },
+    /// A virtual directory error.
+    #[error(transparent)]
+    VirtualDirectoryError(#[from] weaver_common::Error),
 
     /// An invalid registry archive.
     #[error("The registry archive `{archive}` is invalid: {error}")]
