+++ src/registry/diff.rs
@@ -12,11 +12,11 @@ use miette::Diagnostic;
 use serde::Serialize;
 use std::path::PathBuf;
 use weaver_common::diagnostic::{DiagnosticMessage, DiagnosticMessages};
+use weaver_common::vdir::VirtualDirectoryPath;
 use weaver_common::Logger;
 use weaver_forge::config::{Params, WeaverConfig};
 use weaver_forge::file_loader::EmbeddedFileLoader;
 use weaver_forge::{OutputDirective, TemplateEngine};
-use weaver_semconv::registry_path::RegistryPath;
 use weaver_semconv::registry_repo::RegistryRepo;
 
 /// Embedded default schema changes templates
@@ -31,7 +31,7 @@ pub struct RegistryDiffArgs {
 
     /// Parameters to specify the baseline semantic convention registry
     #[arg(long)]
-    baseline_registry: RegistryPath,
+    baseline_registry: VirtualDirectoryPath,
 
     /// Format used to render the schema changes. Predefined formats are: ansi, json,
     /// and markdown.
@@ -159,7 +159,7 @@ mod tests {
     use crate::cli::{Cli, Commands};
     use crate::registry::diff::RegistryDiffArgs;
     use crate::registry::{
-        semconv_registry, RegistryArgs, RegistryCommand, RegistryPath, RegistrySubCommand,
+        semconv_registry, RegistryArgs, RegistryCommand, RegistrySubCommand, VirtualDirectoryPath,
     };
     use crate::run_command;
     use std::fs::OpenOptions;
@@ -177,12 +177,12 @@ mod tests {
             command: Some(Commands::Registry(RegistryCommand {
                 command: RegistrySubCommand::Diff(RegistryDiffArgs {
                     registry: RegistryArgs {
-                        registry: RegistryPath::LocalFolder {
+                        registry: VirtualDirectoryPath::LocalFolder {
                             path: "tests/diff/registry_head/".to_owned(),
                         },
                         follow_symlinks: false,
                     },
-                    baseline_registry: RegistryPath::LocalFolder {
+                    baseline_registry: VirtualDirectoryPath::LocalFolder {
                         path: "tests/diff/registry_baseline/".to_owned(),
                     },
                     diff_format: "json".to_owned(),
@@ -206,12 +206,12 @@ mod tests {
         let registry_cmd = RegistryCommand {
             command: RegistrySubCommand::Diff(RegistryDiffArgs {
                 registry: RegistryArgs {
-                    registry: RegistryPath::LocalFolder {
+                    registry: VirtualDirectoryPath::LocalFolder {
                         path: "tests/diff/registry_head/".to_owned(),
                     },
                     follow_symlinks: false,
                 },
-                baseline_registry: RegistryPath::LocalFolder {
+                baseline_registry: VirtualDirectoryPath::LocalFolder {
                     path: "tests/diff/registry_baseline/".to_owned(),
                 },
                 diff_format: "json".to_owned(),
