+++ src/registry/generate.rs
@@ -158,14 +158,12 @@ mod tests {
 
     use tempdir::TempDir;
 
-    use weaver_common::TestLogger;
-
     use crate::cli::{Cli, Commands};
     use crate::registry::generate::RegistryGenerateArgs;
-    use crate::registry::{
-        PolicyArgs, RegistryArgs, RegistryCommand, RegistryPath, RegistrySubCommand,
-    };
+    use crate::registry::{PolicyArgs, RegistryArgs, RegistryCommand, RegistrySubCommand};
     use crate::run_command;
+    use weaver_common::vdir::VirtualDirectoryPath;
+    use weaver_common::TestLogger;
 
     #[test]
     fn test_registry_generate() {
@@ -186,7 +184,7 @@ mod tests {
                     param: None,
                     params: None,
                     registry: RegistryArgs {
-                        registry: RegistryPath::LocalFolder {
+                        registry: VirtualDirectoryPath::LocalFolder {
                             path: "crates/weaver_codegen_test/semconv_registry/".to_owned(),
                         },
                         follow_symlinks: false,
@@ -261,7 +259,7 @@ mod tests {
                     param: None,
                     params: None,
                     registry: RegistryArgs {
-                        registry: RegistryPath::LocalFolder {
+                        registry: VirtualDirectoryPath::LocalFolder {
                             path: "crates/weaver_codegen_test/semconv_registry/".to_owned(),
                         },
                         follow_symlinks: false,
@@ -308,7 +306,7 @@ mod tests {
                     param: None,
                     params: None,
                     registry: RegistryArgs {
-                        registry: RegistryPath::LocalFolder {
+                        registry: VirtualDirectoryPath::LocalFolder {
                             path: "crates/weaver_codegen_test/semconv_registry/".to_owned(),
                         },
                         follow_symlinks: false,
@@ -413,7 +411,7 @@ mod tests {
                         param: None,
                         params: None,
                         registry: RegistryArgs {
-                            registry: RegistryPath::LocalFolder {
+                            registry: VirtualDirectoryPath::LocalFolder {
                                 path: "data/symbolic_test/".to_owned(),
                             },
                             follow_symlinks,
